Move loginRequired.js from js to app folder
Auditors:

Test Plan: n/a
