+++ js/components/main.js
@@ -33,7 +33,7 @@ const WidevinePanel = require('../../app/renderer/components/widevinePanel')
 const AutofillAddressPanel = require('./autofillAddressPanel')
 const AutofillCreditCardPanel = require('./autofillCreditCardPanel')
 const AddEditBookmark = require('../../app/renderer/components/bookmarks/addEditBookmark')
-const LoginRequired = require('./loginRequired')
+const LoginRequired = require('../../app/renderer/components/loginRequired')
 const ReleaseNotes = require('../../app/renderer/components/releaseNotes')
 const BookmarksToolbar = require('../../app/renderer/components/bookmarks/bookmarksToolbar')
 const ContextMenu = require('./contextMenu')
